Modified Gemfile in Rails app, Added Course Controller
The current Gemfile did not allow me to actually run the rails server,
so I uncommented the "therubyracer" line and added an execjs line. The
.lock file was then genreated by using "bundle install".

I also added an empty course controller and course view, and modified
the home page to be more "railsy" in linking to the course page.