Changed names, added new routes and pages
The route "resources :exercises" was added to the routes file.

The controller exercise_controller was renamed to exercises_controller
to ensure compatibility with the resources route. The corresponding
helper was also renamed.

Methods corresponding to the new routes were added to exercises
controller. A new _form partial was added for making new exercises. This
functionality will be modified in the futute.

Slight modiftications to the application layout and a little bit of css
to the application.css to prevent the navbar from chopping off the top
of the content.