Added Raketask, renamed "order", added exercise_id to scores
A raketask to generate exercise order from completeOrder.txt was
created. WARNING: Old task names were modified.

The attribute "order" in the exercise model was changed to "position"
due to an unavoidable sql syntax error becasue of the name "order"

The exercise controller was changed to order exercises by "position"

The score model has belongs_to exercise, but no exercise_id column, so
this was added.