Added non-working preview functionality
Added file preview.js that contains the code for the dynamic preview.

Placed the testingFunctionParser.html into public. This will cause it to
generate an exercise when running `rake exercise:all` for testing
purposes. This should be removed before merging into master.

The answer-type file was modified so include a place to show the
preview. id-tags were also changed.